Implement x-axis flip for simple TDOS mode
Rather than add arguments to various writers, transform the data once
between plotting and writing.

Ultimately it might be simpler to do it this way everywhere; now we
have a mixture of approaches which is not ideal. But would prefer not
to

- Add extra arguments everywhere
- Break existing API without a major version update.

Maybe cleaning this up should be on the wishlist for Version 2?